# openapi.yaml
openapi: 3.0.1
info:
  title: Ombrella API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
  /eth_getBalance:
    post:
      operationId: eth_getBalance
      summary: Get ETH balance of an address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                tag:
                  type: string
      responses:
        '200':
          description: Returns the balance of the account of a given address.
        '500':
          description: Error while forwarding the request to Alchemy API.
  /eth_getBlockTransactionCountByHash:
    post:
      operationId: eth_getBlockTransactionCountByHash
      summary: Get the number of transactions in a block from a block matching the given block hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hash:
                  type: string
      responses:
        '200':
          description: Returns the number of transactions in a block.
        '500':
          description: Error while forwarding the request to Alchemy API.
  /eth_getBlockTransactionCountByNumber:
    post:
      operationId: eth_getBlockTransactionCountByNumber
      summary: Get the number of transactions in a block matching the given block number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
      responses:
        '200':
          description: Returns the number of transactions in a block.
        '500':
          description: Error while forwarding the request to Alchemy API.
  /eth_blockNumber:
    post:
      operationId: eth_blockNumber
      summary: Returns the number of most recent block
      responses:
        '200':
          description: Returns the number of most recent block.
        '500':
          description: Error while forwarding the request to Alchemy API.
  /eth_getBlockByNumber:
    post:
      operationId: eth_getBlockByNumber
      summary: Returns information about a block by block number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blockNumber:
                  type: string
                returnTransactionObjects:
                  type: boolean
      responses:
        '200':
          description: Returns information about a block.
        '500':
          description: Error while forwarding the request to Alchemy API.
  /eth_getFilterChanges:
    post:
      operationId: eth_getFilterChanges
      summary: Returns an array of all changes for a specific filter since the last time it was called.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filterId:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getLogs:
    post:
      operationId: eth_getLogs
      summary: Returns an array of all logs matching a given filter object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filter:
                  type: object
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getBlockByHash:
    post:
      operationId: eth_getBlockByHash
      summary: Returns information about a block by hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blockHash:
                  type: string
                returnTransactionObjects:
                  type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getTransactionCount:
    post:
      operationId: eth_getTransactionCount
      summary: Returns the number of transactions sent from an address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                tag:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getTransactionReceipt:
    post:
      operationId: eth_getTransactionReceipt
      summary: Returns the receipt of a transaction by transaction hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactionHash:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getTransactionByHash:
    post:
      operationId: eth_getTransactionByHash
      summary: Returns a transaction based on its hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactionHash:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_getTransactionByBlockHashAndIndex:
    post:
      operationId: eth_getTransactionByBlockHashAndIndex
      summary: Returns a transaction based on the block hash and transaction index position.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blockHash:
                  type: string
                transactionIndex:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /eth_call:
    post:
      operationId: eth_call
      summary: Executes a new message call immediately without creating a transaction on the blockchain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                callObject:
                  type: object
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /getTokenAllowance:
    post:
      operationId: getTokenAllowance
      summary: Get the number of tokens that an owner allowed to a spender.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contract:
                  type: string
                owner:
                  type: string
                spender:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Error response
  /getTokenMetadata/{contract}:
    get:
      operationId: getTokenMetadata
      summary: Fetches token metadata for a given contract.
      parameters:
        - in: path
          name: contract
          required: true
          schema:
            type: string
          description: The contract address.
      responses:
        '200':
          description: Successful response with token metadata.
          content:
            application/json:
              schema:
                type: object
                description: The token metadata.
        '500':
          description: Error response
          content:
            text/plain:
              schema:
                type: string
                example: 'Error while fetching token metadata'
  /getNFTsForOwnerV3:
    get:
      operationId: getNFTsForOwnerV3
      summary: Fetches NFTs for a given owner.
      parameters:
        - in: query
          name: owner
          schema:
            type: string
          required: true
          description: The owner address.
        - in: query
          name: withMetadata
          schema:
            type: boolean
          required: false
          description: Whether to include metadata in the response.
        - in: query
          name: pageSize
          schema:
            type: integer
          required: false
          description: The number of items to return per page.
      responses:
        '200':
          description: Successful response with NFTs.
          content:
            application/json:
              schema:
                type: object
                description: The NFTs.
        '500':
          description: Error response
          content:
            text/plain:
              schema:
                type: string
                example: 'Error while forwarding the request to Alchemy API'
  /getNFTMetadata:
    get:
      operationId: getNFTMetadata
      summary: Fetches metadata for a given NFT.
      parameters:
        - in: query
          name: contractAddress
          schema:
            type: string
          required: true
          description: The contract address of the NFT.
        - in: query
          name: tokenId
          schema:
            type: string
          required: true
          description: The token ID of the NFT.
        - in: query
          name: refreshCache
          schema:
            type: boolean
          required: false
          description: Whether to refresh the cache.
      responses:
        '200':
          description: Successful response with NFT metadata.
          content:
            application/json:
              schema:
                type: object
                description: The NFT metadata.
        '500':
          description: Error response
          content:
            text/plain:
              schema:
                type: string
                example: 'Error while forwarding the request to Alchemy API'
...
  /getContractMetadata:
    get:
      operationId: getContractMetadata
      summary: Fetches metadata for a given contract.
      parameters:
        - in: query
          name: contractAddress
          schema:
            type: string
          required: true
          description: The contract address.
      responses:
        '200':
          description: Successful response with contract metadata.
          content:
            application/json:
              schema:
                type: object
                description: The contract metadata.
        '500':
          description: Error response
          content:
            text/plain:
              schema:
                type: string
                example: 'Error while forwarding the request to Alchemy API'
...